update sql table schema

+++ idb/postgres/internal/schema/setup_postgres.sql
@@ -3,7 +3,7 @@
 -- TODO? replace all 'addr bytea' with 'addr_id bigint' and a mapping table? makes addrs an 8 byte int that fits in a register instead of a 32 byte string
 
 CREATE TABLE IF NOT EXISTS block_header (
-  round bigint PRIMARY KEY,
+  round numeric(20) PRIMARY KEY,
   realtime timestamp without time zone NOT NULL,
   rewardslevel bigint NOT NULL,
   header jsonb NOT NULL
@@ -14,10 +14,10 @@ CREATE TABLE IF NOT EXISTS block_header (
 CREATE INDEX IF NOT EXISTS block_header_time ON block_header (realtime);
 
 CREATE TABLE IF NOT EXISTS txn (
-  round bigint NOT NULL,
+  round numeric(20) NOT NULL,
   intra integer NOT NULL,
   typeenum smallint NOT NULL,
-  asset bigint NOT NULL, -- 0=Algos, otherwise AssetIndex
+  asset numeric(20) NOT NULL, -- 0=Algos, otherwise AssetIndex
   txid bytea, -- base32 of [32]byte hash, or NULL for inner transactions.
   txn jsonb NOT NULL, -- json encoding of signed txn with apply data; inner txns exclude nested inner txns
   extra jsonb NOT NULL,
@@ -32,7 +32,7 @@ CREATE INDEX IF NOT EXISTS txn_by_tixid ON txn ( txid );
 
 CREATE TABLE IF NOT EXISTS txn_participation (
   addr bytea NOT NULL,
-  round bigint NOT NULL,
+  round numeric(20) NOT NULL,
   intra integer NOT NULL
 );
 
@@ -46,8 +46,8 @@ CREATE TABLE IF NOT EXISTS account (
   rewardsbase bigint NOT NULL,
   rewards_total bigint NOT NULL,
   deleted bool NOT NULL, -- whether or not it is currently deleted
-  created_at bigint NOT NULL, -- round that the account is first used
-  closed_at bigint, -- round that the account was last closed
+  created_at numeric(20) NOT NULL, -- round that the account is first used
+  closed_at numeric(20), -- round that the account was last closed
   keytype varchar(8), -- "sig", "msig", "lsig", or NULL if unknown
   account_data jsonb NOT NULL -- trimmed ledgercore.AccountData that excludes the fields above; SQL 'NOT NULL' is held though the json string will be "null" iff account is deleted
 );
@@ -55,12 +55,12 @@ CREATE TABLE IF NOT EXISTS account (
 -- data.basics.AccountData Assets[asset id] AssetHolding{}
 CREATE TABLE IF NOT EXISTS account_asset (
   addr bytea NOT NULL, -- [32]byte
-  assetid bigint NOT NULL,
+  assetid numeric(20) NOT NULL,
   amount numeric(20) NOT NULL, -- need the full 18446744073709551615
   frozen boolean NOT NULL,
   deleted bool NOT NULL, -- whether or not it is currently deleted
-  created_at bigint NOT NULL, -- round that the asset was added to an account
-  closed_at bigint, -- round that the asset was last removed from the account
+  created_at numeric(20) NOT NULL, -- round that the asset was added to an account
+  closed_at numeric(20), -- round that the asset was last removed from the account
   PRIMARY KEY (addr, assetid)
 );
 
@@ -72,12 +72,12 @@ CREATE INDEX IF NOT EXISTS account_asset_by_addr_partial ON account_asset(addr) 
 
 -- data.basics.AccountData AssetParams[index] AssetParams{}
 CREATE TABLE IF NOT EXISTS asset (
-  index bigint PRIMARY KEY,
+  index numeric(20) PRIMARY KEY,
   creator_addr bytea NOT NULL,
   params jsonb NOT NULL, -- data.basics.AssetParams; json string "null" iff asset is deleted
   deleted bool NOT NULL, -- whether or not it is currently deleted
-  created_at bigint NOT NULL, -- round that the asset was created
-  closed_at bigint -- round that the asset was closed; cannot be recreated because the index is unique
+  created_at numeric(20) NOT NULL, -- round that the asset was created
+  closed_at numeric(20) -- round that the asset was closed; cannot be recreated because the index is unique
 );
 
 -- For account lookup
@@ -93,12 +93,12 @@ CREATE TABLE IF NOT EXISTS metastate (
 -- per app global state
 -- roughly go-algorand/data/basics/userBalance.go AppParams
 CREATE TABLE IF NOT EXISTS app (
-  index bigint PRIMARY KEY,
+  index numeric(20) PRIMARY KEY,
   creator bytea NOT NULL, -- account address
   params jsonb NOT NULL, -- json string "null" iff app is deleted
   deleted bool NOT NULL, -- whether or not it is currently deleted
-  created_at bigint NOT NULL, -- round that the asset was created
-  closed_at bigint -- round that the app was deleted; cannot be recreated because the index is unique
+  created_at numeric(20) NOT NULL, -- round that the asset was created
+  closed_at numeric(20) -- round that the app was deleted; cannot be recreated because the index is unique
 );
 
 -- For account lookup
@@ -107,11 +107,11 @@ CREATE INDEX IF NOT EXISTS app_by_creator_deleted ON app(creator, deleted);
 -- per-account app local state
 CREATE TABLE IF NOT EXISTS account_app (
   addr bytea,
-  app bigint,
+  app numeric(20),
   localstate jsonb NOT NULL, -- json string "null" iff deleted from the account
   deleted bool NOT NULL, -- whether or not it is currently deleted
-  created_at bigint NOT NULL, -- round that the app was added to an account
-  closed_at bigint, -- round that the account_app was last removed from the account
+  created_at numeric(20) NOT NULL, -- round that the app was added to an account
+  closed_at numeric(20), -- round that the account_app was last removed from the account
   PRIMARY KEY (addr, app)
 );
